Add a mark to an obsplot

Add a mark to an obsplot

mark_area(g, ...)

mark_areaX(g, ...)

mark_areaY(g, ...)

mark_barX(g, ...)

mark_barY(g, ...)

mark_cell(g, ...)

mark_cellX(g, ...)

mark_cellY(g, ...)

mark_dot(g, ...)

mark_dotX(g, ...)

mark_dotY(g, ...)

mark_line(g, ...)

mark_lineX(g, ...)

mark_lineY(g, ...)

mark_link(g, ...)

mark_rect(g, ...)

mark_rectX(g, ...)

mark_rectY(g, ...)

mark_ruleX(g, ...)

mark_ruleY(g, ...)

mark_text(g, ...)

mark_textX(g, ...)

mark_textY(g, ...)

mark_tickX(g, ...)

mark_tickY(g, ...)

Arguments

g
an obsplot object

...
named options passed to define the mark

... arguments can accept at most two unnamed arguments : a data object and a call to a transform function.